A patch from Antoine Vittecoq for "Form widget/ FTL pagination, Last does not take to last page" https://issues.apache.org/jira/browse/OFBIZ-5575
This issue in pagination can be tested in https://localhost:8443/partymgr/control/findparty (FTL) & https://localhost:8443/webtools/control/FindUtilCache?VIEW_SIZE_1=2&VIEW_INDEX_1=0 (XML Form).


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1579277 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/base/src/org/ofbiz/base/util/UtilMisc.java
@@ -790,7 +790,7 @@ public class UtilMisc {
     }
 
     public static int getViewLastIndex(int listSize, int viewSize) {
-        return (int)Math.ceil(listSize / viewSize) - 1;
+        return (int)Math.ceil(listSize / (float) viewSize) - 1;
     }
     
 }
